data-generation/requirements.txt
async-timeout==4.0.3
certifi==2023.7.22
clickhouse-connect==0.6.14
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
importlib-resources==6.0.1
kiwisolver==1.4.5
lz4==4.3.2
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==23.1
Pillow==10.0.0
powerlaw==1.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
scipy==1.10.1
six==1.16.0
urllib3==2.0.6
zipp==3.16.2
zstandard==0.21.0
